Move vault addresses to their own core module
Importing them through core's barrel pulled `knownTokens` and its
`viem/chains` import into every earn entry point, since esbuild
cannot prove the token literal is pure. Earn's shared chunk drops
from 7,339 B to 575 B.
