+++ packages/core/src/index.ts
@@ -1,8 +1,8 @@
 export {
-  knownTokens,
   sVetBtcAddress,
   sVusdAddress,
   stakingVaultAddresses,
-} from "./tokens.ts";
+} from "./stakingVaultAddresses.ts";
+export { knownTokens } from "./tokens.ts";
 export type { Token } from "./types.ts";
 export { isAddressValid } from "./utils/isAddressValid.ts";
