+++ packages/core/src/stakingVaultAddresses.ts
@@ -1,0 +1,9 @@
+import type { Address } from "viem";
+
+export const sVusdAddress: Address =
+  "0x476310E34D2810f7d79C43A74E4D79405bd7a925";
+
+export const sVetBtcAddress: Address =
+  "0x0cB9D84d4bcEc8d3D5B2d99a6F07f4605325987e";
+
+export const stakingVaultAddresses: Address[] = [sVusdAddress, sVetBtcAddress];
