+++ packages/core/src/tokens.ts
@@ -1,15 +1,7 @@
-import type { Address } from "viem";
 import { arbitrum, base, bsc, hemi, mainnet, optimism } from "viem/chains";
 
+import { sVetBtcAddress, sVusdAddress } from "./stakingVaultAddresses.ts";
 import type { Token } from "./types.ts";
-
-export const sVusdAddress: Address =
-  "0x476310E34D2810f7d79C43A74E4D79405bd7a925";
-
-export const sVetBtcAddress: Address =
-  "0x0cB9D84d4bcEc8d3D5B2d99a6F07f4605325987e";
-
-export const stakingVaultAddresses: Address[] = [sVusdAddress, sVetBtcAddress];
 
 export const knownTokens: Token[] = [
   {
